requirements.txt
streamlit
openai
elevenlabs
python-dotenv
